+++ Example.hpp
@@ -50,7 +50,7 @@ class ComputeApp {
         ComputeApp(
             char const* shaderFile,
             uint32_t const bufferSize,
-            float* bufferData,
+            uint32_t* bufferData,
             uint32_t* dims, // [x,y,z],
             uint32_t const* dimLengths // [local_size_x, local_size_y, local_size_z]
         );
@@ -89,7 +89,7 @@ namespace Utility {
     void fillBuffer(
         VkDevice const& device,
         VkDeviceMemory& bufferMemory,
-        float*& bufferData, 
+        uint32_t*& bufferData, 
         uint32_t const bufferSize
     );
     // Creates descriptor set layout
